% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{bodyOf}
\alias{bodyOf}
\title{bodyOf}
\usage{
bodyOf(vec_)
}
\arguments{
\item{vec_}{File readed as vectors.}
}
\value{
Character vector, the shortest path of a certain name in a list
}
\description{
Build a nested list recursively by rule_.
}
\keyword{internal}
