% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{buildContent}
\alias{buildContent}
\title{bulidContent}
\usage{
buildContent(selected_, filename_)
}
\arguments{
\item{selected_}{part designated to source}

\item{filename_}{file to source}
}
\value{
content
}
\description{
Return the specified module of the project file that complies with the "Vinnish" standard.
}
