% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{buildFile}
\alias{buildFile}
\title{buildFile}
\usage{
buildFile(filename_, rule_ = NULL)
}
\arguments{
\item{filename_}{File to source.}
}
\value{
Builded nested list containing codes' head and body.
}
\description{
Build a nested list recursively by rule_.
}
