% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{checkFormat}
\alias{checkFormat}
\title{checkFormat}
\usage{
checkFormat(vec_, rule_, checkType_)
}
\arguments{
\item{vec_}{File readed as vectors.}

\item{rule_}{rule.}
}
\description{
Check if the document title levels comply with the rule.
}
\keyword{internal}
