% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{findFile}
\alias{findFile}
\title{findFile}
\usage{
findFile(keyWord_, path_ = ".", pattern_ = ".R$")
}
\arguments{
\item{path_}{}
}
\description{
Find the files containing the keywords and print out their paths.
}
