% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{findWhere}
\alias{findWhere}
\title{findWhere}
\usage{
findWhere(lst_, selected_)
}
\arguments{
\item{selected_}{}
}
\value{
Character vector, the shortest path of a certain name in a list
}
\description{
Flatten the shortest path of a certain name in a list.
}
\keyword{internal}
