% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{+.character}
\alias{+.character}
\title{`+` method for character}
\usage{
\method{+}{character}(e1, e2)
}
\arguments{
\item{e1}{A character string.}

\item{e2}{Another character string.}
}
\value{
A concatenated character string.
}
\description{
Concatenates two character strings.
}
