% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{+.default}
\alias{+.default}
\title{Default `+` method}
\usage{
\method{+}{default}(a, b)
}
\arguments{
\item{a}{The first operand.}

\item{b}{The second operand.}
}
\value{
The sum of `a` and `b`.
}
\description{
Falls back to the base implementation of the `+` operator.
}
