% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{prepareThis}
\alias{prepareThis}
\title{prepareThis}
\usage{
prepareThis(..., filename_ = rstudioapi::getSourceEditorContext()$path)
}
\arguments{
\item{...}{part designated to source}

\item{filename_}{file to source}
}
\value{
Null
}
\description{
Prepare the specified module of the project file that complies with the "Vinnish" standard.
}
