% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{runThese}
\alias{runThese}
\title{runThese}
\usage{
runThese(..., filename_ = rstudioapi::getSourceEditorContext()$path)
}
\arguments{
\item{...}{part designated to source}

\item{filename_}{file to source}
}
\value{
Null
}
\description{
Run the specified module of the project file that complies with the "Vinnish" standard.
}
