% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_lst.list}
\alias{select_lst.list}
\title{Method for select_lst with list}
\usage{
\method{select_lst}{list}(.data, fun_)
}
\arguments{
\item{.data}{A list to be filtered.}

\item{fun_}{A function that returns a single logical value for each element of the list.}
}
\value{
A subset of the list where `fun_` evaluates to `TRUE`.
}
\description{
This method applies a filtering function to a list and returns a subset of the list.
}
