% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_lst}
\alias{select_lst}
\title{Generic select_lst function}
\usage{
select_lst(.data, ...)
}
\arguments{
\item{.data}{The input data, typically a list.}

\item{...}{Additional arguments passed to methods.}
}
\description{
This is a generic function for selecting elements from a list.
}
