% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{sourceWhich}
\alias{sourceWhich}
\title{sourceWhich}
\usage{
sourceWhich(
  selected_,
  filename_ = rstudioapi::getSourceEditorContext()$path,
  env_ = globalenv()
)
}
\arguments{
\item{selected_}{part designated to source}

\item{filename_}{file to source}
}
\value{
Null
}
\description{
Run the specified module of the project file that complies with the "Vinnish" standard.
}
