% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{+.character}
\alias{+.character}
\title{`+` method for character}
\usage{
\method{+}{character}(a, b)
}
\arguments{
\item{a}{A character string.}

\item{b}{Another character string.}
}
\value{
A concatenated character string.
}
\description{
Concatenates two character strings.
}
