% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{+.default}
\alias{+.default}
\title{Default `+` method}
\usage{
\method{+}{default}(e1, e2)
}
\arguments{
\item{e1}{The first operand.}

\item{e2The}{second operand.}
}
\value{
The sum of `e1` and `e2`.
}
\description{
Falls back to the base implementation of the `+` operator.
}
